move integration tests into relevant files
