+++ test/auth_test.go
@@ -1,0 +1,26 @@
+package test
+
+import "testing"
+
+func TestAuth(t *testing.T) {
+	// t.Fatal("Unauthorized clients cannot call any function other than method.auth.")
+}
+
+func TestGoogleAuth(t *testing.T) {
+	// t.Fatal("Google+ first sign-in (registration) failed with valid credentials")
+	// t.Fatal("Google+ second sign-in (regular) failed with valid credentials")
+	// t.Fatal("Google+ sign-in passed with invalid credentials")
+	// t.Fatal("Authentication was not ACKed")
+}
+
+func TestClientCertAuth(t *testing.T) {
+	// t.Fatal("Authentication failed with a valid client certificate")
+	// t.Fatal("Authenticated with invalid/expired client certificate")
+	// t.Fatal("Authentication was not ACKed")
+}
+
+func TestTokenAuth(t *testing.T) {
+	// t.Fatal("Authentication failed with a valid token")
+	// t.Fatal("Authenticated with invalid/expired token")
+	// t.Fatal("Authentication was not ACKed")
+}
