+++ test/receive_test.go
@@ -1,0 +1,8 @@
+package test
+
+import "testing"
+
+func TestReceiveOfflineQueue(t *testing.T) {
+	// t.Fatal("Failed to receive queued messages after coming online")
+	// t.Fatal("Failed to send ACK for received message queue")
+}
