+++ test/send_test.go
@@ -1,0 +1,16 @@
+package test
+
+import "testing"
+
+func TestSendEcho(t *testing.T) {
+	// t.Fatal("Failed to send a message to the echo user")
+	// t.Fatal("Failed to send batch message to the echo user")
+	// t.Fatal("Failed to send large message to the echo user")
+	// t.Fatal("Did not receive ACK for a message sent")
+	// t.Fatal("Failed to receive a response from echo user")
+	// t.Fatal("Could not send an ACK for an incoming message")
+}
+
+func TestPing(t *testing.T) {
+	// t.Fatal("Pong/ACK was not sent for ping")
+}
